+++ lib/water_drop/sync_producer.rb
@@ -17,8 +17,8 @@ module WaterDrop
       return unless WaterDrop.config.deliver
 
       DeliveryBoy.deliver(message, options)
-    rescue Kafka::Error => error
-      graceful_attempt?(attempts_count, message, options, error) ? retry : raise(error)
+    rescue Kafka::Error => e
+      graceful_attempt?(attempts_count, message, options, e) ? retry : raise(e)
     end
   end
 end